+++ src/vec/mod.rs
@@ -11,7 +11,7 @@ use core::{
     slice,
 };
 
-use crate::len_type::{check_capacity_fits, DefaultLenType, LenType};
+use crate::len_type::{check_capacity_fits, LenType};
 use crate::CapacityError;
 
 mod drain;
@@ -255,8 +255,7 @@ pub struct VecInner<T, LenT: LenType, S: VecStorage<T> + ?Sized> {
 /// For uncommmon capacity values, or in generic scenarios, you may have to provide the `LenT` generic yourself.
 ///
 /// This should be the smallest unsigned integer type that your capacity fits in, or `usize` if you don't want to consider this.
-pub type Vec<T, const N: usize, LenT = DefaultLenType<N>> =
-    VecInner<T, LenT, OwnedVecStorage<T, N>>;
+pub type Vec<T, const N: usize, LenT = usize> = VecInner<T, LenT, OwnedVecStorage<T, N>>;
 
 /// A [`Vec`] with dynamic capacity
 ///
@@ -1738,6 +1737,8 @@ mod tests {
 
     use static_assertions::assert_not_impl_any;
 
+    use crate::len_type::ZeroLenType;
+
     use super::{Vec, VecView};
 
     // Ensure a `Vec` containing `!Send` values stays `!Send` itself.
@@ -1806,7 +1807,7 @@ mod tests {
         {
             let v: Vec<Droppable, 2> = Vec::new();
             let v: Box<Vec<Droppable, 2>> = Box::new(v);
-            let mut v: Box<VecView<Droppable, u8>> = v;
+            let mut v: Box<VecView<Droppable, usize>> = v;
             v.push(Droppable::new()).ok().unwrap();
             v.push(Droppable::new()).ok().unwrap();
             assert_eq!(Droppable::count(), 2);
@@ -1819,7 +1820,7 @@ mod tests {
         {
             let v: Vec<Droppable, 2> = Vec::new();
             let v: Box<Vec<Droppable, 2>> = Box::new(v);
-            let mut v: Box<VecView<Droppable, u8>> = v;
+            let mut v: Box<VecView<Droppable, usize>> = v;
             v.push(Droppable::new()).ok().unwrap();
             v.push(Droppable::new()).ok().unwrap();
             assert_eq!(Droppable::count(), 2);
@@ -2183,7 +2184,7 @@ mod tests {
 
     #[test]
     fn zero_capacity() {
-        let mut v: Vec<u8, 0> = Vec::new();
+        let mut v: Vec<u8, 0, ZeroLenType> = Vec::new();
         // Validate capacity
         assert_eq!(v.capacity(), 0);